+++ types/src/main/java/com/cobber/fta/plugins/FreeText.java
@@ -15,6 +15,10 @@
  */
 package com.cobber.fta.plugins;
 
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 
@@ -40,19 +44,66 @@ public class FreeText extends LogicalTypeInfinite {
 	/** The Semantic type for this Plugin. */
 	public static final String SEMANTIC_TYPE = "FREE_TEXT";
 
-	/** The Regular Express for this Semantic type. */
+	/** The Regular Expression for this Semantic type. */
 	private static final String REGEXP = ".+";
 
 	private String regExp = REGEXP;
 
 	private int SAMPLE_COUNT = 100;
 	private LogicalTypeCode logicalFirst;
-	private static String[] samples;
-	private static String[] verbs = { "contemplated", "painted", "saw", "observed", "studied" };
-	private static String[] pronouns;
-	private static String[] nouns = { "banana", "wall", "church", "cathederal", "strawberry", "mango",
-			"bicycle", "motorbike", "car", "raspberry", "train", "plane" };
+	private String[] samples;
 	private TextProcessor processor;
+	private Map<String, SimpleSamples> simpleSamples = new HashMap<>();
+
+	private static String[] de_verbs = { "betrachtet", "gemalt", "sah", "beobachtet", "studiert" };
+	private static String[] de_base_pronouns = { "sie", "er" };
+	private static String[] de_nouns = { "Banane", "Wand", "Kirche", "Kathedrale", "Erdbeere", "mango",
+			"Fahrrad", "Motorrad", "Auto", "Himbeere", "Zug", "Flugzeug" };
+
+	private static String[] en_verbs = { "contemplated", "painted", "saw", "observed", "studied" };
+	private static String[] en_base_pronouns = { "she", "he", "they" };
+	private static String[] en_nouns = { "banana", "wall", "church", "cathederal", "strawberry", "mango",
+			"bicycle", "motorbike", "car", "raspberry", "train", "plane" };
+
+	class SimpleSamples {
+		String language;
+		List<String> verbs;
+		List<String> pronouns;
+		List<String> nouns;
+		String definiteArticle;
+		String indefiniteArticle;
+		String conjunction;
+
+		SimpleSamples(final String language, final List<String> verbs, final List<String> pronouns, final List<String> nouns) {
+			this.language = language;
+			this.verbs = verbs;
+			this.pronouns = pronouns;
+			this.nouns = nouns;
+
+			if ("DE".equals(language)) {
+				definiteArticle = " der ";
+				indefiniteArticle = " ein ";
+				conjunction = " und ";
+			}
+			else {
+				definiteArticle = " the ";
+				indefiniteArticle = " a ";
+				conjunction = " and ";
+			}
+		}
+
+		String getSample() {
+			return pronouns.get(random.nextInt(pronouns.size())) + " " +
+					verbs.get(random.nextInt(verbs.size())) +
+					definiteArticle +
+					nouns.get(random.nextInt(nouns.size())) +
+					conjunction +
+					verbs.get(random.nextInt(verbs.size())) +
+					indefiniteArticle +
+					nouns.get(random.nextInt(nouns.size())) +
+					".";
+		}
+	}
 
 	/**
 	 * Construct a plugin to detect free-form text based on the Plugin Definition.
@@ -66,29 +117,27 @@ public class FreeText extends LogicalTypeInfinite {
 		if (samples != null)
 			return;
 
-		pronouns = new String[100];
-		pronouns[0] = "She";
-		pronouns[1] = "He";
-		pronouns[2] = "They";
+		if (locale == null)
+			locale = Locale.getDefault();
+		String language = this.locale.getLanguage().toUpperCase(Locale.ROOT);
+
+		List<String> pronouns = new ArrayList<>(Arrays.asList("DE".equals(language) ? de_base_pronouns : en_base_pronouns));
 
 		for (int i = 3; i < 100; i++) {
 			final String name = logicalFirst.nextRandom();
-			pronouns[i] = String.valueOf(name.charAt(0));
+			String initialCap = String.valueOf(name.charAt(0));
 			if (name.length() > 1)
-				pronouns[i] += name.substring(1).toLowerCase(Locale.ROOT);
+				initialCap += name.substring(1).toLowerCase(Locale.ROOT);
+			pronouns.add(initialCap);
 		}
+
+		SimpleSamples simpleSample = new SimpleSamples(language, new ArrayList<>(Arrays.asList("DE".equals(language) ? de_verbs : en_verbs)), pronouns,
+				new ArrayList<>(Arrays.asList("DE".equals(language) ? de_nouns : en_nouns)));
+		simpleSamples.put(language, simpleSample);
 
 		samples = new String[SAMPLE_COUNT];
 		for (int i = 0; i < SAMPLE_COUNT; i++) {
-			samples[i] = pronouns[random.nextInt(pronouns.length)] + " " +
-					verbs[random.nextInt(verbs.length)] +
-					" the " +
-					nouns[random.nextInt(nouns.length)] +
-					" and " +
-					verbs[random.nextInt(verbs.length)] +
-					" a " +
-					nouns[random.nextInt(nouns.length)] +
-					".";
+			samples[i] = simpleSample.getSample();
 		}
 	}